requirement.txt
google-generativeai
flask
flask-cors
google-api-python-client